chatbot-backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
openai==1.14.3
python-dotenv==1.0.1
